+++ src/lib/drivers/driver-registry.ts
@@ -3,7 +3,7 @@ import { PaginationModeEnum } from '../enums/pagination-mode.enum';
 import { IPaginationConfig } from '../interfaces/pagination-config.interface';
 import { IRequestStrategy } from '../interfaces/request-strategy.interface';
 import { IResponseStrategy } from '../interfaces/response-strategy.interface';
-import { DrfResponseOptions, JsonApiResponseOptions, NestjsResponseOptions, NestjsxCrudResponseOptions, ResponseOptions, StrapiResponseOptions } from '../models/response-options';
+import { DrfResponseOptions, JsonApiResponseOptions, NestjsResponseOptions, NestjsxCrudResponseOptions, ResponseOptions, SpringResponseOptions, StrapiResponseOptions } from '../models/response-options';
 import { DrfRequestStrategy } from '../strategies/drf-request.strategy';
 import { DrfResponseStrategy } from '../strategies/drf-response.strategy';
 import { JsonApiRequestStrategy } from '../strategies/json-api-request.strategy';
@@ -18,6 +18,8 @@ import { PostgrestRequestStrategy } from '../strategies/postgrest-request.strate
 import { PostgrestResponseStrategy } from '../strategies/postgrest-response.strategy';
 import { SpatieRequestStrategy } from '../strategies/spatie-request.strategy';
 import { SpatieResponseStrategy } from '../strategies/spatie-response.strategy';
+import { SpringRequestStrategy } from '../strategies/spring-request.strategy';
+import { SpringResponseStrategy } from '../strategies/spring-response.strategy';
 import { StrapiRequestStrategy } from '../strategies/strapi-request.strategy';
 import { StrapiResponseStrategy } from '../strategies/strapi-response.strategy';
 
@@ -112,6 +114,12 @@ export const DRIVERS: Record<DriverEnum, IDriverDefinition> = {
     createRequestStrategy: () => new SpatieRequestStrategy(),
     createResponseStrategy: () => new SpatieResponseStrategy(),
     createResponseOptions: (config) => new ResponseOptions(config)
+  },
+
+  [DriverEnum.SPRING]: {
+    createRequestStrategy: () => new SpringRequestStrategy(),
+    createResponseStrategy: () => new SpringResponseStrategy(),
+    createResponseOptions: (config) => new SpringResponseOptions(config)
   },
 
   [DriverEnum.STRAPI]: {